Use Best Compression
Toolkit currently uses the default zlib compression type which is not accurate to LittleBigPlanet, this PR changes it to use compression type 9 (best compression), which is 1:1 with LittleBigPlanet's compressed files, and results in smaller file sizes. 

Not sure if this was intentional or not, but if so, perhaps consider making the compression types configurable from within the GUI instead.